requirements.txt
celery==5.2.3
Flask==2.2.2
werkzeug==2.2.2
Flask-WTF==1.0.0
flower==1.0.0
pytest==7.0.1
redis==4.1.4
requests==2.31.0
psycopg2-binary==2.9.9
flask-jwt-extended==4.6.0
Flask-Cors==4.0.0